creating a different class to serve as VectorResourceImpl
